Background Color and Padding
Body background colour, and therefore the background colour of all the pages that are to be created, has been changed to a minty green colour (#EEFBEE) that I have established in the design process.

Top padding has been deleted from the body element, and left and right paddings have been set to 10rem above 786 px and 3rem below it.