Account/Register Done, Responsiveness Implemented to Navbar and Footer
Account/Register View has been edited to be in accord with the rest of the pages.

Button style has been changed to fit the navbar buttons and the sign off button.

It is approximately aligned to center from all sides. Fully-responsive. However, responsiveness seems to be a bit clunky. Decide if you will do something about this, later on.