Context and Initializer Set
Database context with model classes and a database initializer is set. This and the rest of the database initializing logic will be to address the demo run of the app.

Later on I will be deciding whether a Demo Run is all we need or are we going to need to deal with the users.