Database Context Targeted
I have create a new controller named Demo Controller that will deal with the demo run's functionality. I AM NOT SURE that this is the proper way of dealing with what's at hand but will look for problems I may have with this to decide whether to take it another direction.

I will work on Events, People and Venues pages in their own branches which are yet to be created.