Data Annotations Added to NewEventViewModel
I think it's proper to rename the NewEvetViewModel as EventViewModel. I don't know why I named it like this in the beginning, at all.

ISSUE:
Anyway. We have an issue. List of PersonViewModels and VenueViewModels that supposed to contain all the members of the possible respective lists are getting lost in the process of validation if it fails. I haven't checked if this prevails in the case of success but let's tackle this issue first.

ACTION:
Examine thoroughly the involved models views and controllers and figure out what you're missing!