3 Part Event Creation
I will attempt to deal with the current problems at hand by creating 3 different Views (they might be partial views). One for setting name date and time, another for setting a venue, and a third one for selecting which people to invite.

Keep in mind the utility of ViewModels and don't try to make something that they're not supposed to be out of them.