List<SelectListItem>
I started the development of a solution for adding people onto an event that is being created by the user.

I will have a group of checkboxes, each for a one person, that has the value of the person's ID (as string) and Text of the Person's Name.

I am currently working on the logic that needs to be implemented to utilize the value of the checkboxes to retreive people from the database and populate the new events PeopleInvited field with them.