Form Controls Functional
All the form controls seem to be working properly and all the properties of a new Event instance are being populated without any error.

However I am having an issue saving this instance as a Database entity and save changes. I suspect that I am making a mistake in my attempt to form a many-to-many relationship. I will research a bit about this and return back to the issue at hand.