Not Today!
I have added a custom validation rule to the NewEventViewModel to restrict the user to only be able to create events that are dated from a day after the user's entry.

Before I start working on the creation of person entities I will make some adjustments to the DemoController to run the default seed method only if it's in Demo mode.