As Is
I will for not leave the initialization of the database as it is (DropCreateAlways<T>). At the later stages of development I might seek a solution to only run this when the Run Demo button is clicked. I will now work on the alignment of a few items on the page as well as displaying the people attending on the event cards.