Three Cases
The previously existing two versions of the navbar (one to be displayed if the user is authenticated and one if not) has been changed slightly.

I've also added a third case for if the application is running on Demo mode.

I have implemented something similar for the Footer but will now move to its respective branch to finish it up.