New Footer Design Implemented
We have three additional classes to use with .Footer. Namely HomeFooter, DemoFooter, and UserFooter; corresponding to whether the welcome page is being displayer, or the demo is running or a user has signed in, respectively.

I have changed the design of the HomeFooter a little bit and also deleted the background-color and border-top properties from the .DemoFooter and UserFooter. They will have buttons, similar to the ones we have in the navigation to give the user additional controls depending on their role.