Cancelled
Creation of this page has been cancelled due to information being so little for each entity and therefore it being unnecessary to have a seperate page for each person.

Option to upload images for the people will also be cancelled, and instead we'll display infos of each person on its respective card. I might increase the card width to be able to accommodate this.