Demo Controller Lightened
I haven't started working on the Venues yet. I have removed a lot of the code from the demo controller and instead created functions inside the models that corresponds to the code that's been removed. Now we only call the functions and the code exists inside the models (or ViewModels).