Image Upload and Display for Venues
Now we can upload image during the creation of the venue. I will choose a default thumbnail to be displayed if there is no image uploaded.

Venue cards will now display images only. But when hovered, they will display the information about the venue and the edit button.

Next up is the implementation of the ImagePath field in EditVenue view.