Delete Person Implemented
Now we can delete Persons as well as Venues.

One issue is though, if we delete all the people attending to an event, the event still remains. But I guess this should be the expected result. I need to give the user ability to add people to the an event though.