Issue With Buttons
I'm calling it a day here. Buttons are unclickable for events. The reason is that the left column of the event card falls on top of it. I tried using z-index property to fix this but didn't work.

I think the first thing to try tomorrow is to move the buttons inside the left column.