Minimum Viable Implementation
EventDetails view now displays event details including the people attending in a fashionable manner. However, before moving on to the implementation of control buttons inside this view, I will quickly add an option for adding an image for the event so that we may have better event cards in the events view.