Edit Event Implemented
Now we are able to edit the event Title, Date, Time and Image.

The problem is, though, that we have to update all of the aforementioned even if we only want to update one. I will try to solve this by using partial views and therefore am taking a break here to investigate further into how I might go about this.