Events/Delete Implemented
Now we can delete an event by clicking the Delete Event button that appears at the bottom of the individual event views.

We don't have a seperate view for this functionality. Rather, we click the button, confirm the deletion of the related item by the Confirm modal and then the controller does the work for us and redirect to the Index action where all the remaining events are displayed.