Image Upload According to User
I am calling it a day here. Where I'm at is that I've implemented user funcitonalities for the whole app, in that the data is displayed according to the user ID. I have implemented UserID fields to ViewModels and I have created a database table for all the users that only contains the user ID. Other information about the users are being retrieved from the default user table anyway.

I've had a slight problem updating the Venue thumbnails. I have fixed it but I have the same problem with the Events. So tomorrow, this is where I start.