Venue Deletion Issue Fixed
An error occured when I was working on the custom error pages. When deleting a venue the app was hitting a foreign key constraint. I have fixed it by first deleting the events set at the Venue. I will quickly check whether the same error occurs with people and fix it.