Controllers Clean, Creation Issue Fixed Across Controllers
Controllers are now clean. The duplicate code for deciding whether or not to throw an exception by checking whether the user ID matches the entity's UserID or the entity's user ID is "demo", has been made at least a couple of lines more concise for its each instance. I have added a SUMMARY at the start of each controller, giving further information about this function, because it became fairly compact.