Modals Set, Logic Approach Cancelled
I have set the modals and the app now restricts the deletion of Events that have the UserID property of "demo".

This approach proved problematic however, due to the population logic of the entities.

I will instead, implement a "IsDemoEntity" property to all the entities and then make the edit/delete possible for only those that has this property set to "false".