Demo Event Restriction
Now it is not possible to edit or delete the two demo events. If an attempt is made by clicking the control icons, a modal is displayed with information, and if an attempt is made by manually entering related URL, no change is made and the user is navigated back to Events/Index.